+++ lib/flip-number.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = number => {
+    if (number > 0)
+        return -number;
+    else if (number < 0)
+        return Math.abs(number);
+    else
+        return 0;
+}
