+++ lib/parse-content-to-json.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const icalToJson = require('ical2json');
+
+module.exports = (data) => {
+    return icalToJson.convert(data);
+}
